Helpers updated
