+++ app/controllers/facilities_management/beta/direct_award_contract_controller.rb
@@ -9,6 +9,14 @@ module FacilitiesManagement
         @page_data[:supplier] = 'Cleaning London LTD'
       end
 
+      def review_and_generate_documents
+        @page_data[:procurement_data] = { contract_name: 'School facilities London', supplier: 'Cabinet office', date_offer_expires: DateTime.new(2019, 7, 7, 8, 2, 0).in_time_zone('London'), contract_number: 'RM330-DA2234-2019', contract_value: '£752,026', framework: 'RM3830', sub_lot: 'sub-lot 1a',
+                                          initial_call_off_period: 7, initial_call_off_start_date: Date.new(2019, 11, 1), initial_call_off_end_date: Date.new(2016, 10, 31),
+                                          expiration_date: DateTime.new(2019, 7, 23, 14, 20, 0).in_time_zone('London'), date_contract_received: DateTime.new(2019, 11, 20, 14, 0, 0).in_time_zone('London'), date_responded_to_contract: DateTime.new(2019, 6, 23, 14, 20, 0).in_time_zone('London'), date_contract_signed: DateTime.new(2019, 6, 23, 14, 20, 0).in_time_zone('London'), date_contract_closed: DateTime.new(2019, 11, 22, 14, 37, 0).in_time_zone('London'),
+                                          mobilisation_period: 4, optional_call_off_extensions_1: 1, optional_call_off_extensions_2: 1, optional_call_off_extensions_3: nil, optional_call_off_extensions_4: nil,
+                                          buildings_and_services: [{ building: 'Barton court store', service_codes: [] }, { building: 'CCS London office 5th floor', service_codes: ['C.13', 'C.20', 'N.1'] }, { building: 'Phoenix house', service_codes: [] }, { building: 'Vale court', service_codes: [] }, { building: 'W Cabinet office 3rd floor', service_codes: [] }] }
+      end
+
       private
 
       def set_page_model
@@ -49,6 +57,16 @@ module FacilitiesManagement
             back_url: ccs_patterns_prototypes_path,
             back_text: 'Back',
             page_title: 'Sending the contract',
+            caption1: 'Total facilities management',
+            continuation_text: 'Confirm and send contract to supplier',
+            return_url: ccs_patterns_prototypes_path,
+            return_text: 'Return to procurement dashboard',
+            secondary_text: 'Cancel, return to review your contract'
+          },
+          review_and_generate_documents: {
+            back_url: ccs_patterns_prototypes_path,
+            back_text: 'Back',
+            page_title: 'Review and generate documents',
             caption1: 'Total facilities management',
             continuation_text: 'Confirm and send contract to supplier',
             return_url: ccs_patterns_prototypes_path,
