+++ app/helpers/facilities_management/beta/direct_award_contract_helper.rb
@@ -1,0 +1,3 @@
+module FacilitiesManagement::Beta::DirectAwardContractHelper
+  include FacilitiesManagement::Beta::RequirementsHelper
+end
