+++ app/views/facilities_management/beta/direct_award_contract/_da_procurement_requirements.html.erb
@@ -1,0 +1,94 @@
+<h1 class="govuk-heading-m govuk-!-margin-bottom-0">
+    1. Requirements
+</h1>
+<div class="govuk-!-width-full govuk-!-font-size-19">
+    <dl class="govuk-summary-list govuk-body">
+        <table class="govuk-table govuk-!-margin-bottom-6">
+            <tbody class="govuk-table__body">
+                <tr class="govuk-table__row">
+                    <td class="govuk-table__cell govuk-!-padding-top-0 govuk-!-width-one-third"></td>
+                    <td class="govuk-table__cell govuk-!-padding-top-0 govuk-!-width-two-thirds"></td>
+                </tr>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Supplier</th>
+                    <td class="govuk-table__cell"><%= @page_data[:procurement_data][:supplier] %></td>
+                </tr>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Contract name</th>
+                    <td class="govuk-table__cell"><%= @page_data[:procurement_data][:contract_name] %></td>
+                </tr>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Contract number</th>
+                    <td class="govuk-table__cell"><%= @page_data[:procurement_data][:contract_number] %></td>
+                </tr>
+                    <th class="govuk-table__cell govuk-table__header">Contract value</th>
+                    <td class="govuk-table__cell"><%= @page_data[:procurement_data][:contract_value] %></td>
+                </tr>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Framework, sub-lot</th>
+                    <td class="govuk-table__cell"><%= "#{@page_data[:procurement_data][:framework]}, #{@page_data[:procurement_data][:sub_lot]}" %></td>
+                </tr>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Initial call-off period</th>
+                    <td class="govuk-table__cell">
+                        <%= govuk_details(pluralize(@page_data[:procurement_data][:initial_call_off_period], 'year')) {initial_call_off_period_description} %>
+                    </td>
+                </tr>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Mobilisation period</th>
+                    <td class="govuk-table__cell">
+                    <% if !@page_data[:procurement_data][:mobilisation_period].nil? %>
+                        <%= govuk_details(pluralize(@page_data[:procurement_data][:mobilisation_period], 'week')) { mobilisation_period_description } %>
+                    <% else %>
+                        None
+                    <% end %>
+                    </td>
+                </tr>
+                <% if !@page_data[:procurement_data][:optional_call_off_extensions_1].nil? %>
+                    <tr class="govuk-table__row">
+                        <th class="govuk-table__cell govuk-table__header">Optional call-off extension period 1</th>
+                        <td class="govuk-table__cell">
+                            <%= govuk_details(pluralize(@page_data[:procurement_data][:optional_call_off_extensions_1], 'year')) { extension_one_description } %>
+                        </td>
+                    </tr>
+                <% end %>
+                <% if !@page_data[:procurement_data][:optional_call_off_extensions_2].nil? %>
+                    <tr class="govuk-table__row">
+                        <th class="govuk-table__cell govuk-table__header">Optional call-off extension period 2</th>
+                        <td class="govuk-table__cell">
+                            <%= govuk_details(pluralize(@page_data[:procurement_data][:optional_call_off_extensions_2], 'year')) { extension_two_description } %>
+                        </td>
+                    </tr>
+                <% end %>
+                <% if !@page_data[:procurement_data][:optional_call_off_extensions_3].nil? %>
+                    <tr class="govuk-table__row">
+                        <th class="govuk-table__cell govuk-table__header">Optional call-off extension period 3</th>
+                        <td class="govuk-table__cell">
+                            <%= govuk_details(pluralize(@page_data[:procurement_data][:optional_call_off_extensions_3], 'year')) { extension_three_description } %>
+                        </td>
+                    </tr>
+                <% end %>
+                <% if !@page_data[:procurement_data][:optional_call_off_extensions_4].nil? %>
+                    <tr class="govuk-table__row">
+                        <th class="govuk-table__cell govuk-table__header">Optional call-off extension period 4</th>
+                        <td class="govuk-table__cell">
+                            <%= govuk_details(pluralize(@page_data[:procurement_data][:optional_call_off_extensions_4], 'year')) { extension_four_description } %>
+                        </td>
+                    </tr>
+                <% end %>
+                <tr class="govuk-table__row">
+                    <th class="govuk-table__cell govuk-table__header">Buildings and services</th>
+                    <td class="govuk-table__cell">
+                        <% @page_data[:procurement_data][:buildings_and_services].each do |bs| %>
+                            <%= govuk_details(bs[:building]) do %>
+                                <% bs[:service_codes].each do |s| %>
+                                  <li><%= to_lower_case(FacilitiesManagement::StaticData.work_packages.select{|work_package| work_package['code'] == s}.first['name']) %></li>
+                                <% end %>
+                            <% end %>
+                        <%end%>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
+    </dl>
+</div>
